Build Depends updated for qt5
There is no libqt5-dev package its named qtbase5-dev .The package also depends on qt5-qmake so its also been removed. 
in order to force building with qt5 we will also need to insure the system uses the qt5 tools to build we can do this by installing the qt5-default package however it might be a bit inconvenent for some people who are wanting to use qt4 . this is only an issue if the user has also installed qt4 dev parts.